Use a qml.qrc to list all qml files
We need both CMake and Qt creator to know what all of our qt sources are,
and we need these to not get out of sync as both devs and ui/ux works on
these files.  This set up keeps one single location where we store the
list of qml files that we use and lets Qt creator automatically use this
list without making it _too_ much harder for devs to edit this list.