Add a window to control the bot model
We are only compiling our kaiten bot model for the bot, so when running
from qt creator we only get default property values.  This is not going
to make qt creator builds very useful going forward as more things are
based off of property values.

The bot model window is still pretty stupid -- it just has a couple of
hard coded fields to change the few properties that the UI currently uses.

This changes some of the layout here a bit -- the qml directory is now
ui because it really should be thought of as the entire ui (in particular
all resources live here, not just qml).