Fix a few initialization bugs
The connected signal wasn't getting to kaiten_bot_model because we
were firing it before it was connected, and we weren't actually creating
our jsonrpc callback objects so we were ignoring the results.

Also we remove most of the extra fields that we added for testing because
it would be confusing -- machine name and IP are still useful though.