Add a process submodel with a few properties
Currently just enough properties to tell us whether a process is running,
what process it is, and what step it is on.  I think that we want to be
using enumerations for process types and steps and most of the various
things that kaiten sends us a string for, but we still want the strings
available for things that we aren't going to be doing anything specific
with.  Except maybe that just makes it more annoying to add a new step?

Also it is currently quite annoying to add a new submodel...