First pass at an error screen for prints
This adds failure as a state, sets the print icon up to display this state,
and unconditionally adds the current error code of the process to the status
screen.  This does not change the contents of the print status screen based
on the state of the process since doing that would probably make a giant merge
conflict with existing efforts to do so.