Actually start the ui at boot
I forgot that we kept the systemd service file for the actual ui in the auther
directory.  Also we might as well still install the auther program and a disabled
systemd service file for it so that anyone with ssh access to the machine can
still easily enable automatic authorization remotely (and also we can still
support prototype printers without a working touch screen).