Update Quick Controls to get some popups back
When we updated to Qt 5.10 all of our popups disappeared.  I did find an
online recommendation to add an overlay parent to popup objects in order to
get them to actually show up, and this does appear to work for the popups
that live on the top level window (authentication and firmware updates) but
not on other popups.  Also overlay objects need quick controls 2.3 so I went
ahead and updated that dependency across the entire UI.