First pass at handling firmware updates from USB.
This commit adds functionality to the MoreporkStorage class to parse a USB drive
for valid firmware files. It reads the 'pid' value from the manifest.json file
to determine the files validity. Also implemented are a non blocking Qt copy
QObject and the ability to call brooklyn upload from QML. At the time of
writting this, Qt does not offer a non-blocking copy function. The copy function
that was implemented can be run both in it's own thread or as a local thread
object. In either case it does not block given it calls it's iterate function by
queuing it up within the UI's main thread. With this commit, we create a QThread
and place the ProgressCopy object within it.