Copy in main thread
Originally, it was decided to copy the firmware file in a separate QThread
because the file transfer on the bot being tested was already super slow.
Copying in the main thread (that runs the UI) would mean Qt is handling UI
events and copying in an interlaced fasion. In other words, Qt is scheduling
the copy. Turns out DVT3 solved some ground loop bugs on the UI that drastically
improved USB transfer rates from the AUX board. Given this, I've decided to
remove the QThread and run the copy in the main thread because the code is just
simply easier to read and is a safer option given that with the previous
implementation we are syncing signals between two threads, now an unnecessary
complexity.