show all unverified zip files, add file type filtering, minor bug fix
Previously we were verifying all zip files on the USB drive as we recursed through the USB drive directories. Only the verified files would display. If there were a large number of zip files on the USB drive, this could be a slow process. Now we display all zip files on the display and we verify the file when the user selects it. There is a function for looking for print files and one for looking for firmware files. Both of these are called by the USB drive storage watcher. If we are looking for print files, we don't need to parse out zip files and vice versa. So we now specify via QML what type of file we are looking for and MoreporkStorage only looks for this file type. Lastly, there was a small QML bug (warning) where we were assigning an undefined variable to a disable_button bool.