Add V2 machine, extruder & materials support
This commit includes two things, support for new extruders including
ways to handle them on different printer types since V2 extruders are
not supported on V1 machines and then material support for the new
extruders.

As for handling exruder the 'tool_type_correct' flag sent in system
notification by kaiten determines the "correctness" by looking through
printer_settings.json under 'supported_tool_types' key. Since the UI
just knows whether the attached tool is correct or not and nothing
about the type of mismatch, the messaging to user is solely determined
based on the printer type. It is the machine/kaiten's responsibility
to tell client's about the mismatch.

As for materials handling, the materials valid for a bay/extruder
depends on the extruder attached. The materials are extruder specific
and not machine specific, but the extruders are themselves machine
specific, so practically the materials are also machine specific. e.g.
hot extruders can never be attached to V1 printers so ABS/SR-30 can't
be used on V1 without some warranty voiding creativity, conversely V1
extruders can be attached to hot bots, but they can only print with
PLA/TOUGH/PET/PVA. This logic shouldn't get too broken since an
extruder should always be attached to a printer anytime the user wants
to print or load which are the only times the material checks can block.

BW-4879
https://jira.makerbot.net/browse/BW-4879