Reset the require material acknowledgement UI flag when cancelling load
Whenever loading process starts kaiten sends a material validity check
notification to all clients that they can use as a trigger to display
popups to user or even try to acknowldge the material themselves. This
fixes a bug where the UI flag that gets set when the notficaton is
received (which also opens the acknowledge popup) wasn't being reset
when the loading itself is cancelled. Since the flag maintained it's
state from the previous acknowledge notification it wasn't changing state
in subsequent loadings upon receiving the notification thereby not opening
the acknowledge popup.

Also added text in the mismatch popup reflecting the supported materials
depending on the extruders installed.

BW-4907
https://jira.makerbot.net/browse/BW-4907