Add traslations support
This just adds updated translation files for all languages that the UI can
actually use, but it doesn't include font support for all languages.

Button widths are now dynamic depending on the width of the button label.
Hence the property label_width in RoundedButton component has been
deprecated. There are still some elements for some languages that are too
long that need to be handled with multiple lines which will involve making
button heights (and dimensions of other components) dynamic as well but
this commit just ensures things aren't too odd looking to begin with.