We do actually want a space in Method X
In general, the solution here is to always cleanly delineate strings that are
for display purposes from actual functional strings (ie URLs).  But since this
entire branch is literally only for V2 blocking firmware, we don't really need
either of these global properties and can just hard code everything that is V2
specific.