Add extruder combination mismatch error
This is a UI side feature to prevent users from installing a
support extruder in FRE and then only get blocked at test print
after leveling, calibrating & loading material because it doesn't
correspond to the model extruder (no way to slice for extruder
combo) but is still individually compatible with the support slot
for the machine.

This needs to evolve in some way when single extruder printing
rolls out.

BW-4975
https://makerbot.atlassian.net/browse/BW-4975