Block mid-print loading of mismatching materials
We have a lot fancier methods of directing users to the correct action when loading filament while idle, but there are a number of issues preventing us from reusing them while paused during a print.  Instead we just silently disable the load button whenever we are paused during a print and attempting to load filament that we would not be able to use to start the same print.