Tweak back button in wifi page for correct behavior while in FRE
Have explicit back button action defined for first page of wifi
screen. The first page back button action is defined in the
parent level item in the settings page that holds the wifi page.
It is also needed here for cases when other pages in the wifi
page are swiped to, which makes the current item control to be
transferred to the wifi page item and then going back from the
first page has incorrect behavior.

This is one of the quirks of having nested SwipeViews with the
current back button logic.

BW-4977
https://makerbot.atlassian.net/browse/BW-4977