Block loading without material recognition from nfc teg
This replaces the old style warranty ultimatum(!) language in the
unknown materials popup where the user could still acknowledge the
risk to continue loading, with a hard block. It  happens when the
bay or extruder filamnt switches are triggered during loading and no
material is detected through the NFC tag. Even when this hard block
popup shows up the UI still waits for any possible NFC read updates
from kaiten and automatically instructs kaiten to continue loading
if it does find an approved material, providing some buffer time for
NFC reads assuming the spool waas just placed on the bay or the NFC
hardware isn't ready for some reason. This, along with the fact that
the load process is only cancelled when the user dismisses the unknown
material popup should minimize false negatives (prevent loading). The
material mismatch behavior is unchanged and will still show up when
incorrect materials are detected for a bay.

Note that this change especially assumes that the NFC hardware is
robust and the reads happend reliably even with some leeway from the
UI.