detect linux distribution
The artifact download utility for the UI was hardcoded to pull artifacts for
Ubuntu 16.04. I was receiving an obscure JsonCpp run time error while trying to
run morepork UI. The binaries compiled for Ubuntu16.04 (GCC 5) seem to not be
linking compatible with GCC 9. This commit changes the project to a cmake
project where we now check during compile time what the distribution is. The
changes include a cmake macro file found from the TLM project that has
platform name conversion appended to it to allow getting a correct makerbot
artifactory platform name.