Translations were either incorrect or too wide
http://makerbot.atlassian.net/browse/BW-4933

* Chinese option now states that it is Chinese (Simplified).
* Japanese no longer states "Japanese Person".
* In Spanish, "Opciones Avanzadas" is now narrower as "Mas Opciones",
    and should no longer overlap the "Material" icon label

The other translation "issues" are actually a problem displaying certain
characters in (so far) Chinese and Japanese. For example, the option
"Chinese (Simplified)" will only display as "Chinese Simplified", as it
is not correctly displaying the width-appropriate parentheses.

Similarly, Japanese is made up of a combination of native and borrowed
(from Chinese) characters, and it is the native characters (Kana) that are
not appearing. See also BW-4937.