Reset UI error flags when the error dictionary is None
This fixes a bug that caused the UI to wrongly report that the model
extruder jammed, on all jams, after one occurence of a model extruder
jam. This happened since the model extruder jammed flag (which was set
during a previous model extruder jam) wasn't being reset when the error
went away and all subsequent jams were reported as model extruder jams.
This augments existing code that only cleared the flags when the reported
error code is 0.

BW-5153
https://makerbot.atlassian.net/browse/BW-5153