adjust popup for detected experimental extruders
http://makerbot.atlassian.net/browse/BW-5117

The text for a popup that displays when an experimental extruder
is detected was amended - and strained against existing "whitespace"
settings & limits.

Vertical whitespace around the popup title and body text were
redistributed, and body text was manually re-wrapped, allowing
the expanded text to stay within the bounds of the popup.