Show alert icon by the calibration button until calibration is run
We already have a popup asking the user to calibrate their extruders.
However, this popup can be safely dismissed without leaving a sliver
of trace on the UI. This alert icon is persistent until the user
actually calibrates their extruders.

We do this on the software update button too, to call user attention
to a pending software update.