Add authentication support using one time password
The printer will reach out to a web service and get a one time password(OTP)
to display on the UI. It'll also start polling the web service periodically
to check whether someone used the OTP to add this printer to their makerbot
account. If someone did, the cloud service will return that user's username
and oauth token as a response to the polling call, which the UI will then
pass onto Kaiten. The UI also has a local timer to expire the current OTP
after 5 minutes and request a new one. This was chosen in place of letting
the web service tell the printer when the OTP expires, beacuse the printer's
local time might not be in sync with the web service's time.

Also, unified code on both the UI and backend to use common components and
follow similar naming scheme for both the types of authentication.

BW-5144
https://makerbot.atlassian.net/browse/BW-5144