Add support for magma
BW-5213
http://makerbot.atlassian.net/browse/BW-5213

This supports parsing of a machine type of "magma" from kaiten.  This is mostly
done by replacing checks for "Lava" with checks for "not Fire", since we generally
expect the extruder support of Magma to be the same as Lava.  In a few places I
added magma as a separate case that did the same thing as lava as this potentially
seemed more appropriate, and when using this type for a product name I chose the
placeholder XL.  This also installs lava test prints to magma, just in case anything
we care about might break if there were no test prints.  The test prints should
refuse to print.

I also removed the current attempt at per machine settings files, since it was
completely broken and resulting in only V1 printers even getting a default settings
file.  There are not currently any per machine UI settings so I am not sure what
was even being attempted here.