Ensure current thing file parse is successful
There are three cases where a .makerbot file is parsed by the UI
backend. First when a print file is directly selected from the UI to
be printed. Next, when a print file is transferred from another client
like MB Print in which case the UI looks for the 'print_file_valid'
notification to know that the file is in place in the current thing
directory before parsing the metadata. The third case is when a print
is started from repl on a local file on the printer which is when the
UI has no direct means of knowing when to start parsing the file. The
easiest way was to look for the start of a print process and try to
read the .makerbot in the current thing directory after an arbitrary
time. This method worked fine in prior versions of Qt even though it
was most likley failing gracefully for print process started from
external clients since the file isn't in place when the read happens.

This commit ensures that the meta data read is successful before
attempting to read the print details object for both cases, print
from external client and print from repl. Prints initiated from UI
are unaffected by this change.

BW-5233
https://makerbot.atlassian.net/browse/BW-5233