Disable load button when a mismatching spool is in the filament bay
Since kaiten now loads the material config for the spool on the bay before
starting the load process, it can fail if the material is incompatible. So
the UI shouldn't in that case allow the user to load. There is also an alert
icon on the UI by the material name signalling the mismatch.

The UI can still handle material mismatches in a reversible way with
feedback to the user when they happen while the load process is already
running.