Add nicer UI for print defects selection page; cleanups
The thing that I really pushed for was to build a list of defects and just
send that instead of sending the full dict. even for sucesss feedback(with
all defects marked as false), because at some point we might allow users to
type and send feedback but the anlaytics guy insisted on sending the whole
thing.