Get bot serial during UI startup
The bot serial is currently only used by the UI when requesting a one
time password for cloud authentication. The serial was populated just
before making the request and callback to update it doesn't seem to
be returning in time to update the serial, leading to the placeholder
iserial being sent to cloudauth server,  which itself somehow goes
through with it and successfully sends an otp back to the printer.
Subsequent requests had the serial variable correctly populated and
sent along with the otp request.

Also, the handshake call to update the serial isn't really specific to
cloudauth and the serial isn't going to change on every call for it to
live in the cloud auth workflow.