Use lightweight loading indicator for thumbnail images
The loading indicator for all images in the cloud queue list was the busy
spinner used thrughout the UI. Since it's an animated element and is
rendered multiple times, one for each prints small thumbnail it had a
significant impact on the scrolling performance of the list. This commit
replaces the loading indicator with a lightweight placeholder circle.