Use Flickable to handle opening of the drawers
The back button on the left corner of the screen is overlapped at its left
edge and the top edge by the dragable handles of the drawers above it --
swipe back drawer on the left edge of the display and the top drawer at the
top edge which is active in certain scenarios.

There is a bug in Qt(reported here https://bugreports.qt.io/browse/QTBUG-59141)
that causes the drawers to not forward the touch events to elements underneath
(specifically the area under the dragMargin) So, the parts of the back button
overlapped by the drawers dragMargin areas don't recieve the touches intended
to the button.

This commit completely replaces the back swipe drawer with the Flickable
component to detect swipes to go back. It also adds the Flickable component
to the top bar to detect down swipes to then pull the top drawer down instead
of using the drawers own drag handles. This ensures that the entire back button
area is able to receive touches.

BW-5529
https://makerbot.atlassian.net/browse/BW-5529