Log state changes in QML
I really wanted to create a separate directory to put the logging
components together, but that also meant having to tweak the way
Cmake installs the qml files on the bot or having to have conditinal
imports in QML depending on the running platform. Cmake always looks
like Greek and Latin to me, so ended up not doing it for now.

BW-5493
https://makerbot.atlassian.net/browse/BW-5493