Load the application from the Qt Resource System
The QML files and other resources are already bundled as binary files
within the app executable, but to let Qt resolve the references to the
bundled .qml files the main qml file needs to be referenced with the URI
prefix (qrc://) of the qt resource system when initially loaded.

We dont need to separately copy all the qml files and load the main qml file
from its absolute path. In addition to getting rid of redundant data we're
also able to hide the qml source code (unless someone unpacks the executable).

See more here - https://doc.qt.io/qt-5/qtquick-deployment.html#managing-resource-files-with-the-qt-resource-system