add indicator for language-change-in-progress
BW-4647
http://makerbot.atlassian.net/browse/BW-4647

Preliminary work on an indicator that "a language change is in progress". I reused the current animated "loading" icon, and
named this patch "preliminary" in case this is visually undesireable.

I also added a language-choice-button disabler when that language is currently selected, to prevent needless waiting for the
same language to be loaded again.

In fact, all buttons ought to be disabled once a language change has been initiated, and re-enabled afterwards, but that may
already have been recorded as a ticket elsewhere.