Create two versions of the UI application
We are now adding debug symbols to everything that we compile, even with
all other debugging disabled.  We don't want to put debug symbols into
the copy of the application that actually ends up in the firmware, so we
make a stripped copy of the application to install there.  We also want
an application copy with debug symbols to be installed to staging.
However, the size of both the executable and debug sections of the
application are absolutely dwarfed by the bundled resources, and putting
two copies of these resources into every artifact we build is a bit much.
So we strip out the section that contains the bundled resources from the
debugging copy.

Note that stripping the bundled resources does seem to leave allocated
space for the resources in the file but leaves the contents zeroed out,
so space is only saved when compressing the file into an artifact.  We
could probably find a way to shrink the uncompressed binary as well but
the artifact size is what really matters here.