Add logging to swipeviews
This commit adds the ability to log swipe requests and swipe completions
along with the swipe index. The indices are logged in their string names
as defined in the SwipeIndex enums for each swipeview.

Note that the logging is only for swipeviews that are used as containers
in the UI (the ones for which the swipes happen programatically instead
of using fingers)

BW-5493
https://makerbot.atlassian.net/browse/BW-5493