Redesign the material page
This commit adds a new design for the material page unifying the extruder
and material page from the old design. All the information and logic from
the old design has been ported as-is from the old design, so some text
overlappings and spacing issues might be present.

Summary of changes --

* Add functionality to get supported materials from extruder config sent in extruder chnage notification. This is necessary for automatically
populating the materials list for loading material in sunflower. This will also eliminate the need for maintaining supported materials list
for loading on Method/Method X. But I dont really want to mess around with too many things in this commit, so the method loading code hasn't been
touched at all. The supported materials extracted from the notification havent been hooked up yet on the model side of the UI. They will be done
in the next commit that actually adds the loading workflow flow for sunflower.

* Adjusted the auto layout settings in the text components TextBody and TextSubheader to not automatically fill the entire width of the parent
layout component and instead enable this on a need basis at the place of instantiation.

* Explicitly set the dimensions of the UI components in the redesigned material page and the sub-components used in it. I was initially leaning toward
a responsive design approach but I'm not sure the flexibility it provides is worth the loss of tight control fixed layouts provide, especially since
the new designs in figma are with very specific dimensions and standardized styles that cant be guaranteed with auto-sizing. I want to eventually get
these dimanesions from a settings file though instead of hardcoding them in QML in the future.

* Cleaned up code in FilamentBayForm.qml for bay specific properties. I no longer think the term Filament Bay is the most applicable term for this
component but I haven't changed it for now.

* Added new components -- ExtruderStatus and MaterialStatus that go in the redesigned material page. Tweaked MaterialIcon component to the new design
and it now lives in the MaterialStatus component.

* Retained all the existing display information and logic from the old design.

* Formatting changes in qml.qrc file.

* This commit does not change the loading workflow page to use the newly desgined Layout templates. I figured there are far too many chnages here
already and have saved it for later as that also does not materially change user facing behavior.

BW-5587
https://makerbot.atlassian.net/browse/BW-5587