Use spool journal to track loaded filaments for sunflower
*Use loaded_filaments in system notification to track loaded material

*Get the supported materials through get_extruder_config() and
extruder_changed notification. The former is needed as the extruder
change notification is tied with the calibration check, so for a
fully calibrated printer this notification isnt sent when the printer
boots up, which is why we manually get the extruder configs when
the UI starts up. But in any other case when an extruder is swapped
after the printer is already powered on the extruder_changed
notification is sent as the new extruder is either a different type,
as well as marked as uncalibrated.

BW-5587
https://makerbot.atlassian.net/browse/BW-5587