Get feedback from user for user-cancelled prints as well
Prints cancelled by the user that are reported by users as non-
failure as still reported as failed in the print feedback but the
defects dictionary sent along has the key 'non_failure_other' set
to true, which can be used to isolate these case where the print
attempt was practically a filure for the user but wasn't the fault
of the printer itself.

User-cancelled prints that are reported as failure by the user have
the same feedback workflow as what the failure feedback for a
completed print currently has.

Prints that have failed ie. auto-cancelled due to errors don't prompt
for feedback at all from the user.

BW-5474
https://makerbot.atlassian.net/browse/BW-5474