Move logging to mbcoreutils
BW-5632
http://makerbot.atlassian.net/browse/BW-5632

Qt creator logging should be unchanged.  We minimally change the hook
that forwards qt logging to boost, so we are now reaching a bit into
the internals of this header only library, but I think this is still
a bit cleaner than adding something to the logging API to support this
use case.