Added a check for filament bay before checking spool info.
BW-5656
http://makerbot.atlassian.net/browse/BW-5656

Based on the logic, even if loaded_filaments is "None" then it would display spool info regardless of whether there is actually filament available at the extruder. While this is applicable when reading spool info from the filament bay, this could be visually confusing when loading and unloading filament for sunflower because there is no filament bay.