Add check for bay filament switch to display cut bent filament screen
BW-5668
http://makerbot.atlassian.net/browse/BW-5668

To stop the cut bent filament screen from appearing when it is unnecessary, the change to check the bayFilamentSwitch was added to the base display page.
When unload is cancelled (and process is cancellable) the materialChangeCancelled flag is not reset which causes the UI to be acting differently in the load state.