A lot of material handling changes in one commit
BW-5661
http://makerbot.atlassian.net/browse/BW-5661

So the initial focus of this was on support for the removal of filament bay
info from kaiten on sunflower machines, but the scope did grow a bit -- sorry.

We continue to have bot model entries for everything that we use on fire/lava
but now these have the default bot model values (which generally indicate no
spool being present) instead of the fake spool info that kaiten was previously
sending which indicated that there were spools present.  All references to all
of these bot model values were scanned through to ensure that none of these
fake values were making their way to the UI, and loading/starting a print were
both tested to ensure that these were not affected.

One big thing that we were still using the fake spool info for on sunflower was
material compatibility checks when starting a print.  These checks were a bit
of a mess because we had three different representations of material types
floating around and weird upper/lower case conversion involved in comparing
them.  This has been replaced by two different representations, one explicitly
only for display to the user and one explicitly only for material compatibility
checks.  A consistent naming convention (Material for compatibility checks,
MaterialName for display) was applied (which does mean converting kaiten's
loaded_filament to loadedMaterials).  We also no longer convert from material
spool id number to material name here -- kaiten already does this.

Finally I noticed that console.log does not actually write to the logs when
the UI is running on the printer, and that if we have any QML syntax errors
our application segfaults instead of exiting cleanly.