Changed UI workflow for load/unload during a print
BW-5669
http://makerbot.atlassian.net/browse/BW-5669

When completing unload of filament when print is paused, after unload the UI will go to the Material Page instead of back to the Print Page. When loading filament to the printer (in the middle of a print), if the Material must be chosen, we don't want the Material Selection Page to appear because the user cannot (and should not) select a new material in the middle of a print. So the UI will not go to that page in that case.