Add cap check for 1C extruder when loading ABS-R
This commit stops a user from loading ABS-R on 1C extruder if their extruder
is a lower subtype (0) than what is required/recommended (subtype 2) for
loading ABS-R on 1C. The user can also reprogram their extruder to subtype 2
after confirming that they installed the cap right from the popup.

This commit also changes ABS-R from being outright blocked from loading on
1C subtypes < 2 to allowing it on all 1C subtypes and then checking for the
subtype requirement furthur down the line in the code. The load button is
generally disabled when incompatible materials are used, so changing that
behavior for this one case was more complexity than allowing it to load and
then blocking it later down the line.

BW-5649
https://makerbot.atlassian.net/browse/BW-5649